refactor(report): add report create api

+++ backend/src/main/java/wooteco/prolog/report/application/dto/ReportAbilityRequest.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class ReportAbilityRequest {
 
     private Long abilityId;
-    private Float weight;
+    private int weight;
 
 }
